[KHP-82] feat: add history for preparation and ingredient

+++ database/migrations/2025_07_07_213732_create_ingredients_locations_tables.php
@@ -46,7 +46,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('ingredient_id')->constrained()->onDelete('cascade')->nullable(false);
             $table->foreignId('location_id')->constrained()->onDelete('cascade')->nullable(false);
-            $table->float('quantity')->default(0)->nullable(false);
+            $table->decimal('quantity', 8, 2)->default(0)->nullable(false);
             $table->timestamps();
         });
     }
